feat: mMARCO and itersplit
This commit adds a PyTorch Dataset wrapper for the mMARCO dataset loader
from hugging face.

It's relatively huge – and requires too much memory for my PC. I used it
initially for testing and smaller training runs, although vscode would
slowly crash as the swap fills up.

I tried to make the dataset streamable, so it does not have to be
materialized in its entirety on system memory. I thought I got this
working, but it proved me wrong: somehow the memory is not freed, or
consumed faster than enough can be freed in reasonable time. It almost
crashes after a few seconds.

Only commiting this to have git history...